Move integration tests, add coverage script
